requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_lambda
aws-cdk.aws_events_targets
